Save workday history do sqlite3 database
This commit introduces a database with history log. It's not very stable now because if there is any problem with the database, the app will not start. It can require manual manipulation with the database (presumably deleting it). In future commits this database history feature will be optional so that the app can run without it.